+++ test/CosmosDB.collections.Tests.ps1
@@ -189,11 +189,27 @@ InModuleScope CosmosDB {
                 $newCosmosDbCollectionIncludedPathIndexParameters = @{
                     Kind      = 'Spatial'
                     DataType  = 'Number'
-                    Precision = -1
                 }
 
                 $errorMessage = $($LocalizedData.ErrorNewCollectionIncludedPathIndexInvalidDataType `
                     -f $newCosmosDbCollectionIncludedPathIndexParameters.Kind, $newCosmosDbCollectionIncludedPathIndexParameters.DataType, 'Point, Polygon, LineString')
+
+                { $script:result = New-CosmosDbCollectionIncludedPathIndex @newCosmosDbCollectionIncludedPathIndexParameters } | Should -Throw $errorMessage
+            }
+        }
+
+        Context 'When called with invalid Spatial Precision parameters' {
+            $script:result = $null
+
+            It 'Should throw expected exception' {
+                $newCosmosDbCollectionIncludedPathIndexParameters = @{
+                    Kind      = 'Spatial'
+                    DataType  = 'Point'
+                    Precision = 1
+                }
+
+                $errorMessage = $($LocalizedData.ErrorNewCollectionIncludedPathIndexPrecisionNotSupported `
+                    -f $newCosmosDbCollectionIncludedPathIndexParameters.Kind)
 
                 { $script:result = New-CosmosDbCollectionIncludedPathIndex @newCosmosDbCollectionIncludedPathIndexParameters } | Should -Throw $errorMessage
             }