+++ CHANGELOG.md
@@ -1,9 +1,12 @@
 # Change Log
 
-## Unreleased
+### Unreleased
 
 - Fix creation of spatial index by `New-CosmosDbCollectionIncludedPathIndex`
   so that precision is not used when passing to `New-CosmosDbCollection`.
+- Added support for `-PartitionKey` in `Invoke-CosmosDbStoredProcedure` - See [Issue #116](https://github.com/PlagueHO/CosmosDB/issues/116)
+- Changed -StoredProcedureParameter from string[] to object[] in `Invoke-CosmosDbStoredProcedure` - See [Issue #116](https://github.com/PlagueHO/CosmosDB/issues/116)
+- Updated `Invoke-CosmosDbStoredProcedure` to set `x-ms-documentdb-script-enable-logging: true` header and write stored procedure logs to the Verbose Stream when `-Debug` is set - See [Issue #116](https://github.com/PlagueHO/CosmosDB/issues/116)
 
 ## 2.0.14.439